backend/requirements.txt
alembic==1.13.2
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
flake8==7.0.0
google-generativeai==0.7.2
PyPDF2==3.0.1
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-docx==1.1.0
pytest==8.3.3
requests==2.32.3
SQLAlchemy==2.0.30
